Share render stats.
