+++ crates/action-runner/src/runner.rs
@@ -9,7 +9,7 @@ use moon_action::{
 use moon_error::MoonError;
 use moon_lang::SupportedLanguage;
 use moon_logger::{color, debug, error, trace};
-use moon_terminal::helpers::replace_style_tokens;
+use moon_terminal::{helpers::replace_style_tokens, ExtendedTerm};
 use moon_utils::time;
 use moon_workspace::Workspace;
 use std::sync::Arc;
@@ -225,6 +225,7 @@ impl ActionRunner {
 
     pub fn render_results(&self, results: &ActionResults) -> Result<(), MoonError> {
         let term = Term::buffered_stdout();
+        term.write_line("")?;
 
         for result in results {
             let status = match result.status {
@@ -261,6 +262,79 @@ impl ActionRunner {
             }
         }
 
+        term.write_line("")?;
+        term.flush()?;
+
+        Ok(())
+    }
+
+    pub fn render_stats(&self, results: &ActionResults, compact: bool) -> Result<(), MoonError> {
+        let mut cached_count = 0;
+        let mut pass_count = 0;
+        let mut fail_count = 0;
+        let mut invalid_count = 0;
+
+        for result in results {
+            if let Some(label) = &result.label {
+                if compact && !label.contains("RunTarget") {
+                    continue;
+                }
+            }
+
+            match result.status {
+                ActionStatus::Cached => {
+                    cached_count += 1;
+                    pass_count += 1;
+                }
+                ActionStatus::Passed | ActionStatus::Skipped => {
+                    pass_count += 1;
+                }
+                ActionStatus::Failed | ActionStatus::FailedAndAbort => {
+                    fail_count += 1;
+                }
+                ActionStatus::Invalid => {
+                    invalid_count += 1;
+                }
+                _ => {}
+            }
+        }
+
+        let mut counts_message = vec![];
+
+        if pass_count > 0 {
+            if cached_count > 0 {
+                counts_message.push(color::success(format!(
+                    "{} completed ({} cached)",
+                    pass_count, cached_count
+                )));
+            } else {
+                counts_message.push(color::success(format!("{} completed", pass_count)));
+            }
+        }
+
+        if fail_count > 0 {
+            counts_message.push(color::failure(format!("{} failed", fail_count)));
+        }
+
+        if invalid_count > 0 {
+            counts_message.push(color::invalid(format!("{} invalid", invalid_count)));
+        }
+
+        let term = Term::buffered_stdout();
+        term.write_line("")?;
+
+        let counts_message = counts_message.join(&color::muted(", "));
+        let elapsed_time = time::elapsed(self.get_duration());
+
+        if compact {
+            term.render_entry("Tasks", &counts_message)?;
+            term.render_entry(" Time", &elapsed_time)?;
+        } else {
+            term.render_entry("Actions", &counts_message)?;
+            term.render_entry("   Time", &elapsed_time)?;
+        }
+
+        term.write_line("")?;
         term.flush()?;
 
         Ok(())
